ports/api/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2
jinja2>=3.1,<4
